refactor(oidc): move OIDCConfiguration out of optional-deps module
