+++ pyproject.toml
@@ -68,8 +68,6 @@ doc = [
     "sphinx-notfound-page==1.1.0",
     "sphinx-copybutton==0.5.2",
     "sphinx-design==0.6.1",
-    "requests_auth>=8.0.0,<9.0.0",
-    "keyring>=22,<26",
 ]
 dev-linux = [
     "asgi_gssapi; sys_platform == 'linux'"
