+++ src/ansys/openapi/common/__init__.py
@@ -34,13 +34,8 @@ from ._exceptions import (
     AuthenticationWarning,
     UndefinedObjectWarning,
 )
+from ._oidc_config import OIDCConfiguration
 from ._session import ApiClientFactory, AuthenticationScheme, OIDCSessionBuilder
-
-try:
-    from ._oidc import OIDCConfiguration
-except ImportError:
-    OIDCConfiguration = None  # type: ignore[misc, assignment]
-
 from ._util import SessionConfiguration, generate_user_agent
 
 __all__ = [
