+++ tests/test_missing_imports.py
@@ -45,6 +45,7 @@ class TestMissingExtras:
     base_module_list = [
         "ansys.openapi.common._session",
         "ansys.openapi.common._oidc",
+        "ansys.openapi.common._oidc_config",
         "ansys.openapi.common",
     ]
 
@@ -73,13 +74,14 @@ class TestMissingExtras:
         package_name = get_package_name()
         assert f"`pip install {package_name}[oidc]`" in str(excinfo.value)
 
-    def test_oidc_configuration_is_none_without_oidc_extra(self, mocker):
+    def test_oidc_configuration_is_available_without_oidc_extra(self, mocker):
         self.blocked_import = "requests_auth"
         mocker.patch("builtins.__import__", side_effect=self.mocked_import)
 
         import ansys.openapi.common as common
 
-        assert common.OIDCConfiguration is None
+        config = common.OIDCConfiguration(client_id="client")
+        assert config.client_id == "client"
 
     @pytest.mark.skipif(os.name == "nt", reason="Test only applies to linux")
     def test_create_autologon_on_linux_with_no_extra_throws(self, mocker):
